Reorganize the e2e test code for exercising encrypting with PEM file so the code to generate a PEm file can be reused
